Back Office

+++ database/migrations/2025_08_28_183454_add_barcode_to_rooms_table.php
@@ -4,7 +4,6 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-
 return new class extends Migration
 {
     /**
@@ -15,14 +14,24 @@ return new class extends Migration
     public function up()
     {
         Schema::table('rooms', function (Blueprint $table) {
-            $table->string('barcode')->unique()->nullable();
+            // Cek dulu apakah kolom 'barcode' sudah ada
+            if (!Schema::hasColumn('rooms', 'barcode')) {
+                $table->string('barcode')->nullable();
+            }
         });
     }
 
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
     public function down()
     {
         Schema::table('rooms', function (Blueprint $table) {
-            $table->dropColumn('barcode');
+            if (Schema::hasColumn('rooms', 'barcode')) {
+                $table->dropColumn('barcode');
+            }
         });
     }
 };
